02 Solved C4312 when casting unsigned int to const void*

+++ Jass/src/Platform/OpenGL/OpenGLVertexArray.cpp
@@ -66,7 +66,7 @@ namespace Jass {
 				ToGLenum(element.Type),
 				element.Normalized ? GL_TRUE : GL_FALSE,
 				vertexBuffer->GetLayout().GetStride(),
-				(const void*)element.Offset);
+				(const void*)(uint64_t)element.Offset);
 			index++;
 		}
